'addStorageFileInConfig'
